feat: Add cache mechanism for storing tokens.
To avoid losing the refresh token, access token, realmId and other
needed information, when the token is obtained for the first time
it will be saved into a file.

After this, whenever the server restarts it will first check if there
is a cached version of the tokens and in case there is, it will be
loaded.