feat: Support for signatures in QBO documents.
Quickbook has an Attachable entity that can be associated with
Invoices and Estimates, which are the two types of documents that
this application will use.

These Attachable will be treated as the signature.

In this implementation a note can be added along with the signature to
specify metadata about it, such as the purpose of the signature.