fix: invalid grant when trying to refresh access token
This issue was caused because the application was not updating the
refresh token each time a new access token was required.

As stated here https://help.developer.intuit.com/s/question/0D50f00006Kfd2X/refreshtokenasync-returning-invalidgrant
each time the access token is refreshed Quickbooks will return a refres
token, sometimes this refresh token can be a different one (around every
24hrs, but could vary). So it is recommended to always store the last
refresh token obtained regardless of wether it is a new one or not.