feat: Add creation of payments.
If the payments only use the parameters specified in the jsdoc documentation
above the createPayment function it will create an Unapplied Quickbooks
Payment.

That is, the payment will be attributed to the customer who issued it,
but it is up to the user of Quickbooks online to decide to which
invoice issue this payment.