feat: Integrated salesman action flow into the mobile app.
Next salesman visit will be assigned to the closest order to the
current location of the salesman.

The salesman visit is properly passed through all the components
of the visit flow.

When the visit is completed the state is updated accordingly.

Added a button to go back to homescreen at the end of the salesman flow.